% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiDD.R
\name{chiDD}
\alias{chiDD}
\title{Rchimerism chiDD}
\usage{
chiDD(sdata, markers, profile, ru, rt, rnn, d1nn, d2nn, d1u, d2u, d1t, d2t,
  r)
}
\arguments{
\item{sdata}{Sample data input text file}

\item{markers}{List of locus markers}

\item{profile, ru, rt, rnn, d1nn, d2nn, d1u, d2u, d1t, d2t, r}{Internal variables for matrix operations}
}
\value{
A data frame with donor percentage results, and sample allele matrix
}
\description{
An internal function to determine donor percentages for a single donor
}
