% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSD.R
\name{chiSD}
\alias{chiSD}
\title{Rchimerism chiSD}
\usage{
chiSD(sdata, markers, profile, rt, dt, d, r)
}
\arguments{
\item{sdata}{Sample data input text file}

\item{markers}{List of locus markers}

\item{profile, rt, dt, d, r}{Internal variables for matrix operations}
}
\value{
A data frame with donor percentage results, and sample allele matrix
}
\description{
An internal function to determine donor percentages for a single donor
}
