% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locDD.R
\name{locDD}
\alias{locDD}
\title{Rchimerism locDD}
\usage{
locDD(donor1_data, donor2_data, recipient_data, markers)
}
\arguments{
\item{donor1_data}{Donor 1 data input text file}

\item{donor2_data}{Donor 2 data input text file}

\item{recipient_data}{Recipient data input text file}

\item{markers}{List of locus markers}
}
\value{
Internal variables used by chiDD.R
}
\description{
An internal function to determine informative loci for double donor cases
}
