% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locSD.R
\name{locSD}
\alias{locSD}
\title{Rchimerism locSD}
\usage{
locSD(ddata, rdata, markers)
}
\arguments{
\item{ddata}{Donor data input text file}

\item{rdata}{Recipient data input text file}

\item{markers}{List of locus markers}
}
\value{
Internal variables used by chiSD.R
}
\description{
An internal function to determine informative loci for a single donor
}
