% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rchimerism.R
\name{Rchimerism}
\alias{Rchimerism}
\title{Rchimerism}
\usage{
Rchimerism()
}
\value{
Returns nothing
}
\description{
Use Rchimerism through the shiny interface
}
\details{
This function allows the user to input data files
and view the chimerism percentages along with the input data matrices
through the browser interface, with the options to download the results
}
